Fix isTransient detection on MetadataFactory for production
In production we rely on annotation cache and Doctrine should never use the driver.
Because of changes introduced in https://github.com/doctrine/orm/pull/8285 it started calling isTransient on MetadataFactory class
and its default implementation uses the driver.

In this commit we add custom implementation of MetadataFactory::isTransient to avoid using the driver class.
We also added an implementation of MappingDriver that throws an exception when called. This was done in order to detect that
such calls happen e.g. in acceptance tests.

[MAILPOET-4215]